requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic
pymysql
alembic
python-dotenv